refactor: move request body to dto
