+++ internal/handlers/dto/handler_body.go
@@ -1,4 +1,4 @@
-package domain
+package dto
 
 type ResetPasswordBody struct {
 	Email string `json:"email" validate:"required,email"`
@@ -8,7 +8,7 @@ type ResponseResetPasswordBody struct {
 	Password string `json:"password" validate:"required"`
 }
 
-type UserBody struct {
+type RegisterRequestBody struct {
 	Password string `json:"password" validate:"required"`
 	Email    string `json:"email" validate:"required,email"`
 	UserName string `json:"username" validate:"required"`
