+++ internal/handlers/user_handler.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/PitiNarak/condormhub-backend/internal/core/domain"
 	"github.com/PitiNarak/condormhub-backend/internal/core/ports"
+	"github.com/PitiNarak/condormhub-backend/internal/handlers/dto"
 	"github.com/PitiNarak/condormhub-backend/pkg/error_handler"
 	"github.com/PitiNarak/condormhub-backend/pkg/http_response"
 	"github.com/go-playground/validator"
@@ -20,7 +21,7 @@ func NewUserHandler(UserService ports.UserService) ports.UserHandler {
 }
 
 func (h *UserHandler) Register(c *fiber.Ctx) error {
-	user := new(domain.UserBody)
+	user := new(dto.RegisterRequestBody)
 	err := c.BodyParser(&user)
 	if err != nil {
 		return error_handler.BadRequestError(err, "your request is invalid")
@@ -109,7 +110,7 @@ func (h *UserHandler) VerifyEmail(c *fiber.Ctx) error {
 }
 
 func (h *UserHandler) ResetPasswordCreate(c *fiber.Ctx) error {
-	body := new(domain.ResetPasswordBody)
+	body := new(dto.ResetPasswordBody)
 
 	if err := c.BodyParser(body); err != nil {
 		return error_handler.BadRequestError(err, "your request is invalid")
@@ -129,7 +130,7 @@ func (h *UserHandler) ResetPasswordCreate(c *fiber.Ctx) error {
 }
 
 func (h *UserHandler) ResetPasswordResponse(c *fiber.Ctx) error {
-	body := new(domain.ResponseResetPasswordBody)
+	body := new(dto.ResponseResetPasswordBody)
 
 	if err := c.BodyParser(body); err != nil {
 		return error_handler.BadRequestError(err, "your request is invalid")
